Generate parse tree
This generates the initial parse tree library which leverages
ANTLR parsing. None of this content is hand-authored.
The `go-vet` static analysis tool in golanglint-ci needed to be disabled
in this so that the generated code can be compiled correctly on CI. This
is to due to ANTLR generated code containing a lot of warnings that are
not relevant to the codebase.