Add initial system package
This adds an initial system package for the FHIRPath implementation.
The types are from the N1 specification, and are beginning with an
initial set of types:

* Booleans
* Integers
* Strings

This is the first step in the implementation of the FHIRPath. This
format is not expected to be the final format, since the `fhir` package
is still itself unstable.