Add `resolver` package
FHIRPath R4 and R5 offer optional support for `resolve(...)`, which
must resolve references to an unspecified FHIRPath resource.

This commit adds a package `resolver` which provides such an abstraction,
as well as a test-double package that provides fakes that can be used
for testing. A no-op version is provided as well, which is a "do-nothing"
version.