Add namespace package
FHIRPath is defined in terms of a set FHIR "Model", which is a namespace
containing all types from the base specification. A model can be
replaced at any time with a custom model that extends the base model
with additional types.

To be able to support custom FHIRPath models, this introduces a
namespace package which encapsulates the naming and containment of types
within a given namespace/model. The name may change in the future and
should not (yet) be considered stable.