Add reflect package
FHIRPath defines a "reflection" specification to allow for getting info
on the type of an object, and to allow for dynamic access to fields and
methods. This commit adds the reflect package to the FHIRPath library
which implements part of this functionality.

Additionally, to make it traversable, a `Reflect` Namespace object has
been included as well.